Crate medit2svg

Structs

Path 🔒

Constants

USAGE 🔒

Functions

add_element_and_neighbors_to_path 🔒

Adds an element and its neighbors that have the same ref as path_ref to el_set.

element_lookup 🔒

Returns a function that looks up an element given its ID.

elements 🔒

Returns the list of elements that are interesting.

frontier 🔒

Returns the path that goes around the given set of elements.

main 🔒

merge_paths 🔒

Merge a collection of connected paths into several unconnected paths.

paths 🔒

Returns the list of “blobs”/frontiers/paths of the mesh’s elements that share the same reference.

twobytwo 🔒

write_svg 🔒